Change: Rename Bar module to BarChart corresponding to the components name
Use same name for the module as the component.
