+++ src/web/components/chart/BarChart.tsx
